+++ SamrStudio/Packages/com.samrstudio.core/Runtime/Events/ScriptableObjects/Vectors/Vector2EventSO.cs
@@ -1,0 +1,15 @@
+/*===============================================================
+Product:    Samr Studio Core
+Developer:  Eric Mitchell
+Company:    Samr Studio
+Date:       05/12/2021 16:50
+================================================================*/
+using UnityEngine;
+
+namespace SamrStudio.Core.Events
+{
+    [CreateAssetMenu(menuName = "SamrStudio/Events/Vectors/Vector2 Event")]
+    public class Vector2EventSO : EventBaseSO<Vector2>
+    {
+    }
+}