+++ SamrStudio/Packages/com.samrstudio.core/Runtime/Events/ScriptableObjects/Vectors/Vector3EventSO.cs
@@ -1,0 +1,16 @@
+/*===============================================================
+Product:    Samr Studio Core
+Developer:  Eric Mitchell
+Company:    Samr Studio
+Date:       05/12/2021 16:53
+================================================================*/
+using UnityEngine;
+
+namespace SamrStudio.Core.Events
+{
+    [CreateAssetMenu(menuName = "SamrStudio/Events/Vectors/Vector3 Event")]
+    public class Vector3EventSO : EventBaseSO<Vector3>
+    {
+
+    }
+}