+++ SamrStudio/Packages/com.samrstudio.core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## Version 1.1.0
+
+Added the following scripts
+- Events/Vectors
+  - Vector2EventSO
+  - Vector3EventSO
+  - Vector2IntEventSO
+  - Vector3IntEventSO
+
 ## Version 1.0.0
 
 Added the following scripts